fix: 修复设置工作台模版参数不正确的问题

+++ weixin-java-cp/src/main/java/me/chanjar/weixin/cp/api/impl/WxCpAgentWorkBenchServiceImpl.java
@@ -26,7 +26,7 @@ public class WxCpAgentWorkBenchServiceImpl implements WxCpAgentWorkBenchService 
   }
 
   @Override
-  public String getWorkBenchTemplate(Long agentId) throws WxErrorException {
+  public String getWorkBenchTemplate(Integer agentId) throws WxErrorException {
     final String url = String.format(this.mainService.getWxCpConfigStorage().getApiUrl(WORKBENCH_TEMPLATE_GET));
     JsonObject jsonObject = new JsonObject();
     jsonObject.addProperty("agentid", agentId);
